Create Viewstatus.php
This file checks if the user is logged in using $_SESSION and retrieves their user type and engineer name from the session. It then retrieves the appropriate status data from the database based on the user type and engineer name.

If the user is an admin, it displays a table of all engineer names and statuses. If the user is not an admin, it displays a table of statuses for the engineer name associated with the session.

Note that you'll need to customize the HTML and database queries to fit the specific needs of your application